Add watches before starting syncers at startup.
Another case where moving the ordering to be obvious helps alleviate any
concerns about race conditions. Watches are added before calling start on
SyncTargets. This might create the possibility that a file in a target is
triggered before the target has been started, but this is not possible
because any change events from a watch are necessarily asynchronous, nor
is it harmful if it would somehow happen.